+++ src/pages/joinBoard/JoinBoardDetail.jsx
@@ -1,875 +1,270 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate, useLocation } from 'react-router-dom';
 import {
-    deleteJoinBoard,
-    getJoinBoard,
-    getJoinBoardCard,
-    getJoinBoardCardByTitle,
-    searchJoinBoard,
-    getAllCommentByJoinBoardId,
-    createComment,
-    deleteComment,
-    updateComment
+  deleteJoinBoard,
+  getJoinBoard,
+  getJoinBoardCard,
 } from '../../service/JoinBoardService';
 import {
-    Container,
-    Typography,
-    Box,
-    Card,
-    CardContent,
-    Divider,
-    IconButton,
-    Menu,
-    MenuItem,
-    List,
-    ListItem,
-    ListItemText,
-    TextField,
-    Button
+  Container,
+  Typography,
+  Box,
+  Card,
+  CardContent,
+  IconButton,
+  Menu,
+  MenuItem,
 } from '@mui/material';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import DeleteIcon from '@mui/icons-material/Delete';
-import EditIcon from '@mui/icons-material/Edit';
-import ReplyIcon from '@mui/icons-material/Reply';
 import MainHeader from "../../components/common/MainHeader";
+import CommentsSection from "../../components/joinboard/CommentsSection";
 import "./JoinBoardDetail.css";
 
 function JoinBoardDetail() {
-    const { id } = useParams();
-    const navigate = useNavigate();
-    const location = useLocation();
-    const [post, setPost] = useState(null);
-    const [loading, setLoading] = useState(true);
-    const [anchorEl, setAnchorEl] = useState(null);
-    const [page, setPage] = useState(1);
-    const [comments, setComments] = useState([]);
+  const { id } = useParams();
+  const navigate = useNavigate();
+  const [post, setPost] = useState(null);
+  const [loading, setLoading] = useState(true);
+  const [anchorEl, setAnchorEl] = useState(null);
+  const [page, setPage] = useState(1);
 
-    // 댓글 작성
-    const [newComment, setNewComment] = useState('');
+  // 게시글 상세 불러오기
+  const fetchPostDetail = async () => {
+    try {
+      const data = await getJoinBoard(id);
+      setPost(data);
+      localStorage.setItem('currentJoinBoardDetail', JSON.stringify({
+        id: data.id,
+        page: page,
+      }));
+    } catch (error) {
+      console.error('게시글을 불러오는 데 실패했습니다.', error);
+    } finally {
+      setLoading(false);
+    }
+  };
 
-    // 댓글 수정
-    const [editingCommentId, setEditingCommentId] = useState(null);
-    const [editCommentContent, setEditCommentContent] = useState('');
+  // 게시글 목록 (페이징) 불러오기
+  const fetchPageData = async () => {
+    const savedPage = localStorage.getItem('currentPage');
+    const pageNum = savedPage ? parseInt(savedPage) : 1;
+    try {
+      await getJoinBoardCard(pageNum - 1);
+      setPage(pageNum);
+    } catch (error) {
+      console.error("게시판 목록을 가져오는 데 실패했습니다.", error);
+    }
+  };
 
-    // 대댓글 작성
-    const [replyToId, setReplyToId] = useState(null);
-    const [replyContent, setReplyContent] = useState('');
+  // 게시글 삭제
+  const handleDeleteClick = async () => {
+    try {
+      await deleteJoinBoard(id);
+      alert('게시글이 삭제되었습니다.');
+      navigate('/join-board', { replace: true });
+    } catch (error) {
+      alert('게시글 삭제에 실패하였습니다.');
+      console.error('게시글 삭제 실패:', error);
+    }
+    handleClose();
+  };
 
-    // IME 조합 중인지 여부 (모든 TextField 공용)
-    const [isComposing, setIsComposing] = useState(false);
+  // 모달(메뉴) 열기/닫기
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
 
-    // 대댓글 입력창 포커스 제어용
-    const replyInputRef = useRef(null); // CHANGED: 대댓글 TextField ref
+  // 게시글 수정 이동
+  const handleEditClick = () => {
+    navigate(`/edit-join-board/${id}`, { replace: false });
+    handleClose();
+  };
 
-    // useEffect(() => {
-    //   if (inputRef.current) {
-    //       inputRef.current.focus();
-    //   }
-    // }, [replyContent]);
-    // REMOVED: replyContent 변화 시마다 포커스 → 커서 튐 원인
+  // 초기 데이터 로드 및 페이지 상태 관리
+  useEffect(() => {
+    fetchPostDetail();
+    fetchPageData();
 
-    // CHANGED: replyToId 바뀔 때만 포커스를 주도록 함
-    useEffect(() => {
-        if (replyToId && replyInputRef.current) {
-            replyInputRef.current.focus();
-        }
-    }, [replyToId]);
+    const savedState = localStorage.getItem('joinBoardState');
+    if (savedState) {
+      const parsedState = JSON.parse(savedState);
+      setPage(parsedState.currentPage || 1);
+    } else {
+      setPage(1);
+    }
 
-    // 게시글 상세 불러오기
-    const fetchPostDetail = async () => {
-        try {
-            const data = await getJoinBoard(id);
-            setPost(data);
-            localStorage.setItem('currentJoinBoardDetail', JSON.stringify({
-                id: data.id,
-                page: page,
-            }));
-        } catch (error) {
-            console.error('게시글을 불러오는 데 실패했습니다.', error);
-        } finally {
-            setLoading(false);
-        }
+    const handlePopState = () => {
+      const savedState = localStorage.getItem('joinBoardState');
+      if (savedState) {
+        const parsedState = JSON.parse(savedState);
+        setPage(parsedState.currentPage || 1);
+      }
     };
 
-    // 게시글 목록 (페이징) 불러오기
-    const fetchPageData = async () => {
-        const savedPage = localStorage.getItem('currentPage');
-        const pageNum = savedPage ? parseInt(savedPage) : 1;
-        try {
-            await getJoinBoardCard(pageNum - 1);
-            setPage(pageNum);
-        } catch (error) {
-            console.error("게시판 목록을 가져오는 데 실패했습니다.", error);
-        }
+    window.addEventListener('popstate', handlePopState);
+    return () => {
+      window.removeEventListener('popstate', handlePopState);
     };
+  }, [id]);
 
-    // 댓글 트리 구조화
-    const organizeComments = (commentsArr) => {
-        const commentMap = {};
-        const rootComments = [];
+  if (loading) return <p>Loading...</p>;
+  if (!post) return <p>게시글을 찾을 수 없습니다.</p>;
 
-        commentsArr.forEach(comment => {
-            commentMap[comment.id] = { ...comment, replies: [] };
-        });
-
-        commentsArr.forEach(comment => {
-            if (comment.parentCommentId) {
-                commentMap[comment.parentCommentId]?.replies.push(commentMap[comment.id]);
-            } else {
-                rootComments.push(commentMap[comment.id]);
-            }
-        });
-
-        return rootComments;
-    };
-
-    const fetchComments = async () => {
-        try {
-            const data = await getAllCommentByJoinBoardId(id);
-            setComments(data);
-        } catch (error) {
-            console.error('댓글을 불러오는 데 실패했습니다.', error);
-        }
-    };
-
-    // 게시글 삭제
-    const handleDeleteClick = async () => {
-        try {
-            await deleteJoinBoard(id);
-            alert('게시글이 삭제되었습니다.');
-            navigate('/join-board', { replace: true });
-        } catch (error) {
-            alert('게시글 삭제에 실패하였습니다.');
-            console.error('게시글 삭제 실패:', error);
-        }
-        handleClose();
-    };
-
-    // 모달(메뉴) 열기/닫기
-    const handleClick = (event) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
-
-    // 게시글 수정 이동
-    const handleEditClick = () => {
-        navigate(`/edit-join-board/${id}`, { replace: false });
-        handleClose();
-    };
-
-    // 댓글 작성
-    const handleCommentSubmit = async () => {
-        if (!newComment.trim()) {
-            alert('댓글 내용을 입력해주세요.');
-            return;
-        }
-
-        try {
-            await createComment(id, { content: newComment });
-            setNewComment('');
-            fetchComments();
-        } catch (error) {
-            console.error('댓글 작성 실패:', error);
-        }
-    };
-
-    // 댓글 삭제
-    const handleDeleteComment = async (commentId) => {
-        try {
-            await deleteComment(commentId);
-            fetchComments();
-        } catch (error) {
-            console.error('댓글 삭제 실패:', error);
-        }
-    };
-
-    // 댓글 수정 시작
-    const handleEditStart = (comment) => {
-        setEditingCommentId(comment.id);
-        // 기존 내용 세팅
-        setEditCommentContent(comment.content);
-    };
-
-    // 댓글 수정 취소
-    const handleEditCancel = () => {
-        setEditingCommentId(null);
-        setEditCommentContent('');
-    };
-
-    // 댓글 수정 제출
-    const handleEditSubmit = async (commentId) => {
-        if (!editCommentContent.trim()) {
-            alert('댓글 내용을 입력해주세요.');
-            return;
-        }
-        try {
-            await updateComment(commentId, { content: editCommentContent });
-            setEditingCommentId(null);
-            setEditCommentContent('');
-            fetchComments();
-        } catch (error) {
-            console.error('댓글 수정 실패:', error);
-            alert('댓글 수정에 실패했습니다.');
-        }
-    };
-
-    // 대댓글 작성 시작
-    const handleReplyStart = (commentId) => {
-        setReplyToId(commentId);
-        setReplyContent('');
-    };
-
-    // 대댓글 작성 취소
-    const handleReplyCancel = () => {
-        setReplyToId(null);
-        setReplyContent('');
-    };
-
-    // 대댓글 제출
-    const handleReplySubmit = async (parentCommentId) => {
-        if (!replyContent.trim()) {
-            alert('댓글 내용을 입력해주세요.');
-            return;
-        }
-
-        try {
-            await createComment(
-                id,
-                { content: replyContent },
-                parentCommentId
-            );
-            setReplyToId(null);
-            setReplyContent('');
-            fetchComments();
-        } catch (error) {
-            console.error('댓글 작성 실패:', error);
-        }
-    };
-
-    // 초기 데이터 로드
-    useEffect(() => {
-        fetchPostDetail();
-        fetchComments();
-
-        const savedState = localStorage.getItem('joinBoardState');
-        if (savedState) {
-            const parsedState = JSON.parse(savedState);
-            setPage(parsedState.currentPage || 1);
-        } else {
-            setPage(1);
-        }
-
-        const handlePopState = () => {
-            const savedState = localStorage.getItem('joinBoardState');
-            if (savedState) {
-                const parsedState = JSON.parse(savedState);
-                setPage(parsedState.currentPage || 1);
-            }
-        };
-
-        window.addEventListener('popstate', handlePopState);
-        return () => {
-            window.removeEventListener('popstate', handlePopState);
-        };
-    }, [id]);
-
-    // 다른 저장된 상태 로드
-    useEffect(() => {
-        const savedDetail = localStorage.getItem('currentJoinBoardDetail');
-        if (savedDetail) {
-            const parsedDetail = JSON.parse(savedDetail);
-            if (parsedDetail.id === parseInt(id)) {
-                setPage(parsedDetail.page);
-            }
-        }
-    }, [id]);
-
-    // 댓글 렌더링
-    const renderComments = () => {
-        const organizedComments = organizeComments(comments);
-        if (organizedComments.length === 0) {
-            return (
-                <Typography variant="body2" sx={{ color: '#777' }}>
-                    댓글이 없습니다.
-                </Typography>
-            );
-        }
-        return organizedComments.map(comment => (
-            <CommentItem key={comment.id} comment={comment} />
-        ));
-    };
-
-    // 댓글/대댓글 Item
-    const CommentItem = ({ comment, depth = 0 }) => {
-        return (
-            <>
-                <ListItem
-                    sx={{
-                        display: 'flex',
-                        justifyContent: 'space-between',
-                        borderBottom: '1px solid #eee',
-                        paddingY: 2,
-                        marginLeft: `${depth * 40}px`,
-                        flexDirection:
-                            editingCommentId === comment.id ? 'column' : 'row',
-                        backgroundColor: depth > 0 ? '#f9f9f9' : 'transparent'
-                    }}
-                >
-                    {editingCommentId === comment.id ? (
-                        // 댓글 수정 모드
-                        <Box
-                            sx={{
-                                width: '100%',
-                                display: 'flex',
-                                flexDirection: 'column',
-                                gap: 2
-                            }}
-                        >
-                            <TextField
-                                fullWidth
-                                value={editCommentContent}
-                                // CHANGED: IME 처리
-                                onChange={(e) => {
-                                    if (!isComposing) {
-                                        setEditCommentContent(e.target.value);
-                                    }
-                                }}
-                                onCompositionStart={() => setIsComposing(true)}
-                                onCompositionEnd={(e) => {
-                                    setIsComposing(false);
-                                    setEditCommentContent(e.target.value);
-                                }}
-                                variant="outlined"
-                                multiline
-                                // autoFocus 제거 또는 필요 시 조건부
-                                InputProps={{
-                                    sx: {
-                                        '&:hover .MuiOutlinedInput-notchedOutline': {
-                                            borderColor: 'black'
-                                        },
-                                        '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                                            borderColor: 'black'
-                                        }
-                                    }
-                                }}
-                            />
-                            <Box
-                                sx={{
-                                    display: 'flex',
-                                    gap: 1,
-                                    justifyContent: 'flex-end'
-                                }}
-                            >
-                                <Button
-                                    className="comment-cancel-update-button"
-                                    variant="contained"
-                                    onClick={handleEditCancel}
-                                >
-                                    취소
-                                </Button>
-                                <Button
-                                    className="comment-update-button"
-                                    variant="contained"
-                                    onClick={() => handleEditSubmit(comment.id)}
-                                >
-                                    수정 완료
-                                </Button>
-                            </Box>
-                        </Box>
-                    ) : (
-                        // 일반 댓글 모드
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: 2,
-                                width: '100%'
-                            }}
-                        >
-                            <Box
-                                sx={{
-                                    display: 'flex',
-                                    flexDirection: 'column',
-                                    alignItems: 'center',
-                                    gap: 1,
-                                    minWidth: '100px'
-                                }}
-                            >
-                                <Box
-                                    sx={{
-                                        width: '40px',
-                                        height: '40px',
-                                        borderRadius: '50%',
-                                        overflow: 'hidden'
-                                    }}
-                                >
-                                    <img
-                                        src={
-                                            comment.memberProfileUrl ||
-                                            'https://www.cheonyu.com/_DATA/product/63900/63992_1672648178.jpg'
-                                        }
-                                        alt={`${comment.memberNickname}의 프로필`}
-                                        style={{
-                                            width: '100%',
-                                            height: '100%',
-                                            objectFit: 'cover'
-                                        }}
-                                    />
-                                </Box>
-                                <Typography
-                                    className="member-nickname"
-                                    sx={{
-                                        fontSize: '0.8rem',
-                                        color: '#333333'
-                                    }}
-                                >
-                                    {comment.memberNickname}
-                                </Typography>
-                            </Box>
-
-                            <ListItemText
-                                primary={comment.content}
-                                secondary={`${comment.createdAt}`}
-                                sx={{
-                                    '& .MuiListItemText-primary': {
-                                        color: '#333',
-                                        fontSize: '1rem',
-                                        marginBottom: '5px'
-                                    },
-                                    '& .MuiListItemText-secondary': {
-                                        color: '#999999',
-                                        fontSize: '0.8rem'
-                                    }
-                                }}
-                            />
-
-                            <Box
-                                sx={{ display: 'flex', gap: 1, ml: 'auto' }}
-                            >
-                                {depth < 1 && (
-                                    // 대댓글 버튼 (1-depth까지만 대댓글 허용)
-                                    <IconButton
-                                        onClick={() =>
-                                            handleReplyStart(comment.id)
-                                        }
-                                        sx={{ color: '#999' }}
-                                    >
-                                        <ReplyIcon />
-                                    </IconButton>
-                                )}
-                                <IconButton
-                                    onClick={() => handleEditStart(comment)}
-                                    sx={{ color: '#999' }}
-                                >
-                                    <EditIcon />
-                                </IconButton>
-                                <IconButton
-                                    edge="end"
-                                    onClick={() =>
-                                        handleDeleteComment(comment.id)
-                                    }
-                                    sx={{ color: '#999' }}
-                                >
-                                    <DeleteIcon />
-                                </IconButton>
-                            </Box>
-                        </Box>
-                    )}
-                </ListItem>
-
-                {/* 대댓글 작성 폼 */}
-                {replyToId === comment.id && (
-                    <Box
-                        sx={{
-                            marginLeft: `${(depth + 1) * 40}px`,
-                            marginTop: 2,
-                            marginBottom: 2,
-                            display: 'flex',
-                            flexDirection: 'column',
-                            gap: 1
-                        }}
-                    >
-                        <TextField
-                            fullWidth
-                            inputRef={replyInputRef} // CHANGED: ref 지정
-                            value={replyContent}
-                            // CHANGED: IME 처리
-                            onChange={(e) => {
-                                if (!isComposing) {
-                                    setReplyContent(e.target.value);
-                                }
-                            }}
-                            onCompositionStart={() => setIsComposing(true)}
-                            onCompositionEnd={(e) => {
-                                setIsComposing(false);
-                                setReplyContent(e.target.value);
-                            }}
-                            placeholder="답글을 입력하세요"
-                            variant="outlined"
-                            size="small"
-                            // autoFocus 제거 (대댓글 열릴 때만 포커스)
-                            InputProps={{
-                                sx: {
-                                    '&:hover .MuiOutlinedInput-notchedOutline': {
-                                        borderColor: 'black'
-                                    },
-                                    '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                                        borderColor: 'black'
-                                    }
-                                }
-                            }}
-                        />
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                gap: 1,
-                                justifyContent: 'flex-end'
-                            }}
-                        >
-                            <Button
-                                className="comment-cancel-update-button"
-                                variant="contained"
-                                onClick={handleReplyCancel}
-                            >
-                                취소
-                            </Button>
-                            <Button
-                                className="comment-update-button"
-                                variant="contained"
-                                onClick={() =>
-                                    handleReplySubmit(comment.id)
-                                }
-                            >
-                                답글 작성
-                            </Button>
-                        </Box>
-                    </Box>
-                )}
-
-                {/* 재귀적으로 대댓글(답글) 렌더링 */}
-                {comment.replies?.map((reply) => (
-                    <CommentItem
-                        key={reply.id}
-                        comment={reply}
-                        depth={depth + 1}
-                    />
-                ))}
-            </>
-        );
-    };
-
-    if (loading) return <p>Loading...</p>;
-    if (!post) return <p>게시글을 찾을 수 없습니다.</p>;
-
-    return (
-        <>
-            <MainHeader />
-            <Container
-                maxWidth="lg"
-                sx={{
-                    padding: '40px 20px',
-                    backgroundColor: '#fff',
-                    minHeight: '100vh'
-                }}
+  return (
+    <>
+      <MainHeader />
+      <Container
+        maxWidth="lg"
+        sx={{
+          padding: '40px 20px',
+          backgroundColor: '#fff',
+          minHeight: '100vh',
+        }}
+      >
+        <Card
+          sx={{
+            backgroundColor: '#fff',
+            borderRadius: '8px',
+            boxShadow: 3,
+            padding: '20px',
+            position: 'relative',
+          }}
+        >
+          <CardContent>
+            {/* 게시글 상세 내용 */}
+            <IconButton
+              sx={{
+                position: 'absolute',
+                top: '10px',
+                right: '10px',
+              }}
+              onClick={handleClick}
             >
-                <Card
-                    sx={{
-                        backgroundColor: '#fff',
-                        borderRadius: '8px',
-                        boxShadow: 3,
-                        padding: '20px',
-                        position: 'relative'
-                    }}
+              <MoreVertIcon />
+            </IconButton>
+            <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} onClose={handleClose}>
+              <MenuItem onClick={handleEditClick}>수정</MenuItem>
+              <MenuItem onClick={handleDeleteClick}>삭제</MenuItem>
+            </Menu>
+
+            <Box
+              sx={{
+                display: 'flex',
+                alignItems: 'flex-start',
+                gap: '20px',
+                marginTop: '50px',
+                marginLeft: '20px',
+              }}
+            >
+              {/* 프로필 섹션 */}
+              <Box
+                sx={{
+                  display: 'flex',
+                  flexDirection: 'column',
+                  alignItems: 'center',
+                  width: '70px',
+                }}
+              >
+                <Box
+                  sx={{
+                    width: '70px',
+                    height: '70px',
+                    borderRadius: '50%',
+                    overflow: 'hidden',
+                    marginBottom: '10px',
+                  }}
                 >
-                    <CardContent>
-                        <IconButton
-                            sx={{
-                                position: 'absolute',
-                                top: '10px',
-                                right: '10px'
-                            }}
-                            onClick={handleClick}
-                        >
-                            <MoreVertIcon />
-                        </IconButton>
-                        <Menu
-                            anchorEl={anchorEl}
-                            open={Boolean(anchorEl)}
-                            onClose={handleClose}
-                        >
-                            <MenuItem onClick={handleEditClick}>수정</MenuItem>
-                            <MenuItem onClick={handleDeleteClick}>삭제</MenuItem>
-                        </Menu>
+                  <img
+                    src={
+                      post.memberProfileUrl ||
+                      'https://www.cheonyu.com/_DATA/product/63900/63992_1672648178.jpg'
+                    }
+                    alt={`${post.memberNickname}의 프로필`}
+                    style={{
+                      width: '100%',
+                      height: '100%',
+                      objectFit: 'cover',
+                      borderRadius: '50%',
+                    }}
+                  />
+                </Box>
+                <Typography
+                  className="member-nickname"
+                  sx={{
+                    fontSize: '0.9rem',
+                    textAlign: 'center',
+                    color: '#333333',
+                    width: '100px',
+                  }}
+                >
+                  {post.memberNickname}
+                </Typography>
+              </Box>
 
-                        {/* 프로필섹션, 제목, 작성, 수정일자 컨테이너 */}
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'flex-start',
-                                gap: '20px',
-                                boxSizing: 'border-box',
-                                marginTop: '50px',
-                                marginLeft: '20px'
-                            }}
-                        >
-                            {/* 프로필 섹션 */}
-                            <Box
-                                sx={{
-                                    display: 'flex',
-                                    flexDirection: 'column',
-                                    alignItems: 'center',
-                                    width: '70px',
-                                    boxSizing: 'border-box',
-                                    marginTop: '0px'
-                                }}
-                            >
-                                <Box
-                                    sx={{
-                                        width: '70px',
-                                        height: '70px',
-                                        borderRadius: '50%',
-                                        overflow: 'hidden',
-                                        marginBottom: '10px'
-                                    }}
-                                >
-                                    <img
-                                        src={
-                                            post.memberProfileUrl ||
-                                            'https://www.cheonyu.com/_DATA/product/63900/63992_1672648178.jpg'
-                                        }
-                                        alt={`${post.memberNickname}의 프로필`}
-                                        style={{
-                                            width: '100%',
-                                            height: '100%',
-                                            objectFit: 'cover',
-                                            aspectRatio: '1/1',
-                                            borderRadius: '50%'
-                                        }}
-                                    />
-                                </Box>
-                                <Typography
-                                    className="member-nickname"
-                                    sx={{
-                                        fontSize: '0.9rem',
-                                        textAlign: 'center',
-                                        color: '#333333',
-                                        width: '100px'
-                                    }}
-                                >
-                                    {post.memberNickname}
-                                </Typography>
-                            </Box>
+              {/* 타이틀 및 작성 정보 */}
+              <Box
+                sx={{
+                  flexGrow: 1,
+                  marginLeft: '40px',
+                  marginTop: '5px',
+                }}
+              >
+                <Typography
+                  sx={{
+                    fontWeight: 'bold',
+                    fontSize: '2rem',
+                    color: '#333',
+                    marginBottom: '10px',
+                  }}
+                >
+                  {post.title}
+                </Typography>
+                <Typography variant="body2" sx={{ color: '#999999' }}>
+                  작성: {post.createdAt}
+                </Typography>
+                <Typography variant="body2" sx={{ color: '#999999' }}>
+                  수정: {post.updatedAt}
+                </Typography>
+              </Box>
+            </Box>
 
-                            {/* 타이틀 섹션 */}
-                            <Box
-                                sx={{
-                                    flexGrow: 1,
-                                    boxSizing: 'border-box',
-                                    marginLeft: '40px',
-                                    marginTop: '5px'
-                                }}
-                            >
-                                <Typography
-                                    sx={{
-                                        display: 'flex',
-                                        alignItems: 'center',
-                                        gap: '10px',
-                                        marginBottom: '10px'
-                                    }}
-                                >
-                                    <Box
-                                        sx={{
-                                            display: 'flex',
-                                            flexDirection: 'column',
-                                            alignItems: 'left'
-                                        }}
-                                    >
-                                        <Box>
-                                            <Typography
-                                                sx={{
-                                                    fontWeight: 'bold',
-                                                    fontSize: '2rem',
-                                                    color: '#333',
-                                                    marginBottom: '10px'
-                                                }}
-                                            >
-                                                {post.title}
-                                            </Typography>
-                                        </Box>
+            {/* 게시글 본문 (예시) */}
+            <Box sx={{ marginTop: '40px', marginLeft: '50px' }}>
+              <Typography variant="h6" sx={{ color: '#333' }}>
+                주제: {post.topic}
+              </Typography>
+              <Typography variant="h6" sx={{ color: '#333', mt: 1 }}>
+                팀 이름: {post.teamName}
+              </Typography>
+              <Typography variant="h6" sx={{ color: '#333', mt: 1 }}>
+                소개글: {post.projectBio}
+              </Typography>
+              <Typography variant="h6" sx={{ color: '#333', mt: 1 }}>
+                팀원 소개: {post.teamBio}
+              </Typography>
+              <Typography variant="h6" sx={{ color: '#333', mt: 1 }}>
+                프로젝트 기간: {new Date(post.startDate).toLocaleDateString()} ~ {new Date(post.endDate).toLocaleDateString()}
+              </Typography>
+              <Typography variant="h6" sx={{ color: '#333', mt: 1 }}>
+                현재 팀 인원: {post.peopleNumber}명
+              </Typography>
+              <Box sx={{ mt: 4 }}>
+                <Typography sx={{ color: '#333', fontSize: '18px', whiteSpace: 'pre-line' }}>
+                  {post.content}
+                </Typography>
+              </Box>
+            </Box>
 
-                                        <Box sx={{ boxSizing: 'border-box' }}>
-                                            <Typography
-                                                variant="body2"
-                                                sx={{
-                                                    color: '#999999',
-                                                    boxSizing: 'border-box',
-                                                    paddingBottom: '3px',
-                                                    marginBottom: '0px'
-                                                }}
-                                            >
-                                                작성: {post.createdAt}
-                                            </Typography>
-
-                                            <Typography
-                                                variant="body2"
-                                                sx={{
-                                                    color: '#999999',
-                                                    boxSizing: 'border-box',
-                                                    paddingTop: '3px',
-                                                    marginTop: '0px'
-                                                }}
-                                            >
-                                                수정: {post.updatedAt}
-                                            </Typography>
-                                        </Box>
-                                    </Box>
-                                </Typography>
-                            </Box>
-                        </Box>
-
-
-                        {/* 나머지 본문 섹션 */}
-                        {/*<Divider sx={{ marginY: 2, backgroundColor: '#ddd' }} />*/}
-                        <Box sx={{ marginTop: '100px' }}>
-                            {/* 주제 */}
-                            <Box sx={{ display: 'flex', alignItems:'center', boxSizing: 'border-box' , padding: '0px' , marginBottom: '40px', marginLeft: '50px' }}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999'}}>
-                                    <strong>주제</strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px' , display: 'inline', padding: '0px' }}>
-                                    {post.topic}
-                                </Typography>
-                            </Box>
-
-                            {/* 팀 이름 */}
-                            <Box sx={{ display: 'flex',  alignItems:'center', marginTop: 1 , boxSizing: 'border-box' , padding: '0px' , marginBottom: '40px', marginLeft: '50px'}}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999' }}>
-                                    <strong>팀 이름</strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px' ,padding: '0px' , marginBottom: '0px'}}>
-                                    {post.teamName}
-                                </Typography>
-                            </Box>
-
-                            {/* 소개글 */}
-                            <Box sx={{ display: 'flex', alignItems:'center', marginTop: 1, boxSizing: 'border-box', padding: '0px' , marginBottom: '40px', marginLeft: '50px' }}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999' }}>
-                                    <strong>소개글</strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px', padding: '0px' , marginBottom: '20px' }}>
-                                    {post.projectBio}
-                                </Typography>
-                            </Box>
-
-                            {/* 팀원 소개 */}
-                            <Box sx={{ display: 'flex', alignItems:'center', marginTop: 1,  boxSizing: 'border-box', padding: '0px'  , marginBottom: '40px', marginLeft: '50px'}}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999' }}>
-                                    <strong>팀원 소개</strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px', padding: '0px' , marginBottom: '0px' }}>
-                                    {post.teamBio}
-                                </Typography>
-                            </Box>
-
-
-                            {/* 프로젝트 기간 (시작일~종료일) */}
-                            <Box sx={{ display: 'flex', alignItems:'center', marginTop: 1, boxSizing: 'border-box', padding: '0px'  , marginBottom: '40px', marginLeft: '50px' }}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999' }}>
-                                    <strong> 프로젝트 기간 </strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px', padding: '0px' , marginBottom: '0px' }}>
-                                    {new Date(post.startDate).toLocaleDateString()} ~ {new Date(post.endDate).toLocaleDateString()}
-                                </Typography>
-                            </Box>
-
-
-
-                            <Box sx={{ display: 'flex', alignItems:'center', marginTop: 1, boxSizing: 'border-box', padding: '0px'  , marginBottom: '40px', marginLeft: '50px' }}>
-                                <Typography sx={{ fontWeight: 600, color: '#999999' }}>
-                                    <strong>현재 팀 인원</strong>
-                                </Typography>
-                                <Typography sx={{ marginLeft: '10px', color: '#333', fontSize: '18px', padding: '0px' , marginBottom: '0px' }}>
-                                    {post.peopleNumber}명
-                                </Typography>
-                            </Box>
-
-
-
-                            {/* 내용 */}
-                            <Box sx={{ display: 'flex', flexDirection: 'column', marginTop:'200px', boxSizing: 'border-box', paddingTop:'0px', marginBottom: '350px',
-                                marginLeft: '50px'}}>
-
-                                <Typography sx={{ boxSizing: 'border-box',  color: '#333', fontSize: '18px', whiteSpace: 'pre-line' }}>
-                                    {post.content}
-                                </Typography>
-                            </Box>
-                        </Box>
-
-
-
-                        {/* 댓글 섹션 수정 */}
-                        <Box sx={{ marginTop: '40px', boxSizing: 'border-box',
-                            padding: '0px',
-                            marginLeft: '0px'}}>
-                            <Divider sx={{ marginY: 2, backgroundColor: '#ddd' }} />
-                            <Typography variant="h6" sx={{ color: '#333' }}></Typography>
-
-                            <List>
-                                {renderComments()}
-                            </List>
-
-
-                            {/* 댓글 입력란 */}
-                            <Box sx={{ display: 'flex', flexDirection: 'column', marginTop: '20px' }}>
-                                <TextField
-                                    variant="outlined"
-                                    label="댓글을 작성하세요"
-                                    value={newComment}
-                                    onChange={(e) => {
-                                        if (!isComposing) {
-                                            setNewComment(e.target.value);}
-                                    }}
-                                    onCompositionStart={() => setIsComposing(true)}
-                                    onCompositionEnd={(e) => {
-                                        setIsComposing(false);
-                                        setNewComment(e.target.value);
-                                    }}
-                                    sx={{ marginBottom: '10px' , "& .MuiOutlinedInput-root": {
-                                            "&:hover .MuiOutlinedInput-notchedOutline": {
-                                                borderColor: "black", // 마우스 호버 시 테두리 검정색
-                                            },
-                                            "&.Mui-focused .MuiOutlinedInput-notchedOutline": {
-                                                borderColor: "black", // 클릭(포커스) 시 테두리 검정색
-                                            },
-                                        }, '& .MuiInputLabel-outlined': {
-                                            color: "gray",
-                                            '&.Mui-focused': {
-                                                color: 'black',
-                                            }},
-
-
-                                    }}
-
-                                />
-
-                                <Button
-                                    className="comment-create-button"
-                                    variant="contained"
-                                    color="black"
-                                    onClick={handleCommentSubmit}
-                                    sx={{ alignSelf: 'flex-end' }}
-
-                                >
-                                    댓글 작성
-                                </Button>
-                            </Box>
-                        </Box>
-                    </CardContent>
-                </Card>
-            </Container>
-        </>
-    );
+            {/* 댓글 섹션 컴포넌트 */}
+            <CommentsSection joinBoardId={id} />
+          </CardContent>
+        </Card>
+      </Container>
+    </>
+  );
 }
+
 export default JoinBoardDetail;
